switching from LaTeX to Typst
Typst is cool and feels kinda neat, compared to the very obtuse and old school LaTeX and its extremely inconsistent documentation. Seriously, I shouldn't have to consult 10 different websites to learn how to use basic features of the language. Typst has very basic documentation (e.g. paragraphs are underlyingly blocks and can be edited that way too, this is not reflected in the docs) but it's open source, has garnered a lot of enthusiasm and contributors in very short time, and even if it falls apart I'll always have LaTeX to fall back on.